Рефакторинг

+++ Domain/ContractsRealization/DebtPaymentService.cs
@@ -3,7 +3,7 @@ using CommunalServices.Domain.Entities;
 
 namespace CommunalServices.Domain.ContractsRealization
 {
-    public class DebtPaymentService(IRepository _repository): IDebtPaymentService
+    public class DebtPaymentService(IRepository _repository, IBankPaymentService _bankPaymentService): IDebtPaymentService
     {
         public async Task<Debt> PayDebtAsync(int debtId)
         {
@@ -13,9 +13,10 @@ namespace CommunalServices.Domain.ContractsRealization
                 return null;
 
             var paymentAccount = await _repository.AddPaymentAccountAsync(new PaymentAccount { DebtId = debtId });
+            bool isSuccessPayment = await _bankPaymentService.ProcessPaymentOperationAsync(paymentAccount);
 
-            var payOperation = new MockBankOperation();
-            payOperation.SendPaymentAccount(paymentAccount);
+            if (!isSuccessPayment)
+                return null;
 
             await _repository.RemovePaymentAccountAsync(paymentAccount.PaymentId);
             await _repository.RemoveDebtAsync(debt.DebtId);
